Generate full test and types
