+++ src/lib.rs
@@ -1,9 +1,6 @@
 use parcel::{parsers::byte::expect_byte, ParseResult, Parser};
 use std::fmt::Debug;
 
-#[cfg(test)]
-mod tests;
-
 /// CycleCost represents an object that can take n cycles to process.
 /// typically this will be assigned to an instruction.
 pub trait CycleCost {
@@ -21,7 +18,9 @@ pub trait ByteSized {
 }
 
 pub mod addressing_mode;
+use addressing_mode::*;
 pub mod mnemonic;
+use mnemonic::*;
 
 /// Instruction represents a single mos6502 instruction, taking a mnemonic and address mode as parameters.
 #[derive(Clone, Copy, Debug, PartialEq)]
@@ -57,661 +56,203 @@ where
     }
 }
 
-macro_rules! gen_instruction_cycles_and_parser {
-    ($mnemonic:ty, $addressing_mode:ty, $opcode:literal, $cycles:literal) => {
-        impl CycleCost for Instruction<$mnemonic, $addressing_mode> {
-            fn cycles(&self) -> usize {
-                $cycles
+macro_rules! generate_instructions {
+    ($($name:ident: $mnemonic:tt, $addressing_mode:tt, $opcode:expr, $cycles:expr,)*) => {
+        $(
+            impl CycleCost for Instruction<$mnemonic, $addressing_mode> {
+                fn cycles(&self) -> usize {
+                    $cycles
+                }
             }
-        }
 
-        impl<'a> Parser<'a, &'a [u8], Instruction<$mnemonic, $addressing_mode>>
-            for Instruction<$mnemonic, $addressing_mode>
-        {
-            fn parse(
-                &self,
-                input: &'a [u8],
-            ) -> ParseResult<&'a [u8], Instruction<$mnemonic, $addressing_mode>> {
-                // If the expected opcode and addressing mode match, map it to a
-                // corresponding Instruction.
-                parcel::map(
-                    parcel::and_then(expect_byte($opcode), |_| <$addressing_mode>::default()),
-                    |am| Instruction::new(<$mnemonic>::default(), am),
-                )
-                .parse(input)
+            impl<'a> Parser<'a, &'a [u8], Instruction<$mnemonic, $addressing_mode>>
+                for Instruction<$mnemonic, $addressing_mode>
+            {
+                fn parse(
+                    &self,
+                    input: &'a [u8],
+                ) -> ParseResult<&'a [u8], Instruction<$mnemonic, $addressing_mode>> {
+                    // If the expected opcode and addressing mode match, map it to a
+                    // corresponding Instruction.
+                    parcel::map(
+                        parcel::and_then(expect_byte($opcode), |_| <$addressing_mode>::default()),
+                        |am| Instruction::new(<$mnemonic>::default(), am),
+                    )
+                    .parse(input)
+                }
             }
+        )*
+
+        #[cfg(test)]
+        mod tests {
+            use crate::addressing_mode::*;
+            use crate::mnemonic::*;
+            use parcel::prelude::v1::Parser;
+            $(
+                #[test]
+                fn $name() {
+                    let bytecode: [u8; 3] = [$opcode, 0x00, 0x00];
+                    assert_eq!(
+                        $crate::Instruction::new(<$mnemonic>::default(), <$addressing_mode>::default()),
+                        $crate::Instruction::new(<$mnemonic>::default(), <$addressing_mode>::default()).parse(&bytecode).unwrap().unwrap()
+                    )
+                }
+            )*
         }
     };
 }
 
-// Arithmetic Operations
-
-// ADC
-
-gen_instruction_cycles_and_parser!(mnemonic::ADC, addressing_mode::Absolute, 0x6d, 4);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ADC,
-    addressing_mode::AbsoluteIndexedWithX,
-    0x7d,
-    4
+#[rustfmt::skip]
+generate_instructions!(
+    adc_absolute: ADC, Absolute, 0x6d, 4,
+    adc_absolute_indexed_with_x: ADC, AbsoluteIndexedWithX, 0x7d, 4,
+    adc_absolute_indexed_with_y: ADC, AbsoluteIndexedWithY, 0x79, 4,
+    adc_indirect_y_indexed: ADC, IndirectYIndexed, 0x71, 5,
+    adc_immediate: ADC, Immediate, 0x69, 2,
+    adc_x_indexed_indirect: ADC, XIndexedIndirect, 0x61, 6,
+    adc_zeropage: ADC, ZeroPage, 0x65, 3,
+    adc_zeropage_indexed_with_x: ADC, ZeroPageIndexedWithX, 0x75, 4,
+    sbc_absolute: SBC, Absolute, 0xed, 4,
+    sbc_absolute_indexed_with_x: SBC, AbsoluteIndexedWithX, 0xFD, 4,
+    sbc_absolute_indexed_with_y: SBC, AbsoluteIndexedWithY, 0xF9, 4,
+    sbc_indirect_y_indexed: SBC, IndirectYIndexed, 0xf1, 5,
+    sbc_immediate: SBC, Immediate, 0xe9, 2,
+    sbc_x_indexed_indirect: SBC, XIndexedIndirect, 0xe1, 6,
+    sbc_zeropage: SBC, ZeroPage, 0xe5, 3,
+    sbc_zeropage_indexed_with_x: SBC, ZeroPageIndexedWithX, 0xf5, 4,
+    and_absolute: AND, Absolute, 0x2d, 4,
+    and_absolute_indexed_with_x: AND, AbsoluteIndexedWithX, 0x3d, 4,
+    and_absolute_indexed_with_y: AND, AbsoluteIndexedWithY, 0x39, 4,
+    and_indirect_y_indexed: AND, IndirectYIndexed, 0x31, 5,
+    and_immediate: AND, Immediate, 0x29, 2,
+    and_x_indexed_indirect: AND, XIndexedIndirect, 0x21, 6,
+    and_zeropage: AND, ZeroPage, 0x25, 3,
+    and_zeropage_indxed_with_x: AND, ZeroPageIndexedWithX, 0x35, 4,
+    asl_absolute: ASL, Absolute, 0x0e, 6,
+    asl_absolute_indexed_with_x: ASL, AbsoluteIndexedWithX, 0x1e, 7,
+    asl_accumulator: ASL, Accumulator, 0x0a, 2,
+    asl_zeropage: ASL, ZeroPage, 0x06, 5,
+    asl_zeropage_indexed_with_x: ASL, ZeroPageIndexedWithX, 0x16, 6,
+    bit_absolute: BIT, Absolute, 0x2c, 4,
+    bit_zeropage: BIT, ZeroPage, 0x24, 3,
+    eor_absolute: EOR, Absolute, 0x4d, 4,
+    eor_absolute_indexed_with_x: EOR, AbsoluteIndexedWithX, 0x5d, 4,
+    eor_absolute_indexed_with_y: EOR, AbsoluteIndexedWithY, 0x59, 4,
+    eor_indirect_y_indexed: EOR, IndirectYIndexed, 0x51, 5,
+    eor_immediate: EOR, Immediate, 0x49, 2,
+    eor_x_indexed_indirect: EOR, XIndexedIndirect, 0x41, 6,
+    eor_zeropage: EOR, ZeroPage, 0x45, 3,
+    eor_zeropage_indexed_with_x: EOR, ZeroPageIndexedWithX, 0x55, 4,
+    lsr_absolute: LSR, Absolute, 0x4e, 6,
+    lsr_absolute_indexed_with_x: LSR, AbsoluteIndexedWithX, 0x5e, 7,
+    lsr_accumulator: LSR, Accumulator, 0x4a, 2,
+    lsr_zeropage: LSR, ZeroPage, 0x46, 5,
+    lsr_zeropage_indexed_with_x: LSR, ZeroPageIndexedWithX, 0x56, 6,
+    ora_absolute: ORA, Absolute, 0x0d, 4,
+    ora_absolute_indexed_with_x: ORA, AbsoluteIndexedWithX, 0x1d, 4,
+    ora_absolute_indexed_with_y: ORA, AbsoluteIndexedWithY, 0x19, 4,
+    ora_indirect_y_indexed: ORA, IndirectYIndexed, 0x11, 5,
+    ora_immediate: ORA, Immediate, 0x09, 2,
+    ora_x_indexed_indirect: ORA, XIndexedIndirect, 0x01, 6,
+    ora_zeroapage: ORA, ZeroPage, 0x05, 3,
+    ora_zeropage_indexed_with_x: ORA, ZeroPageIndexedWithX, 0x15, 4,
+    rol_absolute: ROL, Absolute, 0x2e, 6,
+    rol_absolute_indexed_with_x: ROL, AbsoluteIndexedWithX, 0x3e, 7,
+    rol_accumulator: ROL, Accumulator, 0x2a, 2,
+    rol_zeropage: ROL, ZeroPage, 0x26, 5,
+    rol_zeropage_indexed_with_x: ROL, ZeroPageIndexedWithX, 0x36, 6,
+    ror_absolute: ROR, Absolute, 0x6e, 6,
+    ror_absolute_indexed_with_x: ROR, AbsoluteIndexedWithX, 0x7e, 7,
+    ror_acumulator: ROR, Accumulator, 0x6a, 2,
+    ror_zeropage: ROR, ZeroPage, 0x66, 5,
+    ror_zeropage_indexed_with_x: ROR, ZeroPageIndexedWithX, 0x76, 6,
+    bcc_relative: BCC, Relative, 0x90, 2,
+    bcs_relative: BCS, Relative, 0xb0, 2,
+    beq_relative: BEQ, Relative, 0xf0, 2,
+    bmi_relative: BMI, Relative, 0x30, 2,
+    bne_relative: BNE, Relative, 0xd0, 2,
+    bpl_relative: BPL, Relative, 0x10, 2,
+    bvc_relative: BVC, Relative, 0x50, 2,
+    bvs_relative: BVS, Relative, 0x70, 2,
+    clc_implied: CLC, Implied, 0x18, 2,
+    cld_implied: CLD, Implied, 0xd8, 2,
+    cli_implied: CLI, Implied, 0x58, 2,
+    clv_implied: CLV, Implied, 0xb8, 2,
+    cmp_absolute: CMP, Absolute, 0xcd, 4,
+    cmp_absolute_indexed_with_x: CMP, AbsoluteIndexedWithX, 0xdd, 4,
+    cmp_absolute_indexed_with_y: CMP, AbsoluteIndexedWithY, 0xd9, 4,
+    cmp_indirect_y_indexed: CMP, IndirectYIndexed, 0xd1, 5,
+    cmp_immediate: CMP, Immediate, 0xc9, 2,
+    cmp_x_indexed_indirect: CMP, XIndexedIndirect, 0xc1, 6,
+    cmp_zeropage: CMP, ZeroPage, 0xc5, 3,
+    cmp_zeropage_indexexd_with_x: CMP, ZeroPageIndexedWithX, 0xd5, 4,
+    cpx_absolute: CPX, Absolute, 0xec, 4,
+    cpx_immediate: CPX, Immediate, 0xe0, 2,
+    cpx_zeroage: CPX, ZeroPage, 0xe4, 3,
+    cpy_absolute: CPY, Absolute, 0xcc, 4,
+    cpy_immediate: CPY, Immediate, 0xc0, 2,
+    cpy_zeropage: CPY, ZeroPage, 0xc4, 3,
+    dec_absolute: DEC, Absolute, 0xce, 6,
+    dec_absolute_indexed_with_x: DEC, AbsoluteIndexedWithX, 0xde, 7,
+    dec_zeropage: DEC, ZeroPage, 0xc6, 5,
+    dec_zeropage_indexed_with_x: DEC, ZeroPageIndexedWithX, 0xd6, 6,
+    dex_implied: DEX, Implied, 0xca, 2,
+    dey_implied: DEY, Implied, 0x88, 2,
+    inc_absolute: INC, Absolute, 0xee, 6,
+    inc_absolute_indexed_with_x: INC, AbsoluteIndexedWithX, 0xfe, 7,
+    inc_zeropage: INC, ZeroPage, 0xe6, 5,
+    inc_zeropage_indexed_with_x: INC, ZeroPageIndexedWithX, 0xf6, 6,
+    inx_implied: INX, Implied, 0xe8, 2,
+    iny_implied: INY, Implied, 0xc8, 2,
+    jmp_absolute: JMP, Absolute, 0x4c, 3,
+    jmp_indirect: JMP, Indirect, 0x6c, 5,
+    jsr_absolute: JSR, Absolute, 0x20, 6,
+    lda_immediate: LDA, Immediate, 0xa9, 2,
+    lda_zeropage: LDA, ZeroPage, 0xa5, 3,
+    lda_zeropage_indexed_with_x: LDA, ZeroPageIndexedWithX, 0xb5, 4,
+    lda_absolute: LDA, Absolute, 0xad, 4,
+    lda_absolute_indexed_with_x: LDA, AbsoluteIndexedWithX, 0xbd, 4,
+    lda_absolute_indexed_with_y: LDA, AbsoluteIndexedWithY, 0xb9, 4,
+    lda_indirect_y_indexed: LDA, IndirectYIndexed, 0xb1, 5,
+    lda_x_indexed_indirect: LDA, XIndexedIndirect, 0xa1, 6,
+    ldx_absolute: LDX, Absolute, 0xae, 4,
+    ldx_absolute_indexed_with_y: LDX, AbsoluteIndexedWithY, 0xbe, 4,
+    ldx_immediate: LDX, Immediate, 0xa2, 2,
+    ldx_zeropage: LDX, ZeroPage, 0xa6, 3,
+    ldx_zeropage_indexed_with_y: LDX, ZeroPageIndexedWithY, 0xb6, 4,
+    ldy_absolute: LDY, Absolute, 0xac, 4,
+    ldy_absolute_indexed_with_x: LDY, AbsoluteIndexedWithX, 0xbc, 4,
+    ldy_immediate: LDY, Immediate, 0xa0, 2,
+    ldy_zeropage: LDY, ZeroPage, 0xa4, 3,
+    ldy_zeropage_indexed_with_x: LDY, ZeroPageIndexedWithX, 0xb4, 4,
+    pha_implied: PHA, Implied, 0x48, 3,
+    php_implied: PHP, Implied, 0x08, 3,
+    pla_implied: PLA, Implied, 0x68, 4,
+    plp_implied: PLP, Implied, 0x28, 4,
+    rti_implied: RTI, Implied, 0x40, 6,
+    rts_implied: RTS, Implied, 0x60, 6,
+    sec_implied: SEC, Implied, 0x38, 2,
+    sed_implied: SED, Implied, 0xf8, 2,
+    sei_implied: SEI, Implied, 0x78, 2,
+    sta_absolute: STA, Absolute, 0x8d, 4,
+    sta_absolute_indexed_with_x: STA, AbsoluteIndexedWithX, 0x9d, 5,
+    sta_absolute_indexed_with_y: STA, AbsoluteIndexedWithY, 0x99, 5,
+    sta_indirect_y_indexed: STA, IndirectYIndexed, 0x91, 6,
+    sta_x_indexed_indirect: STA, XIndexedIndirect, 0x81, 6,
+    sta_zeropage: STA, ZeroPage, 0x85, 3,
+    sta_zeropage_indexed_with_x: STA, ZeroPageIndexedWithX, 0x95, 4,
+    stx_absolute: STX, Absolute, 0x8e, 4,
+    stx_zeropage: STX, ZeroPage, 0x86, 3,
+    stx_zeropage_indexed_with_y: STX, ZeroPageIndexedWithY, 0x96, 4,
+    sty_absolute: STY, Absolute, 0x8c, 4,
+    sty_zeropage: STY, ZeroPage, 0x84, 3,
+    sty_zeropage_indexed_with_x: STY, ZeroPageIndexedWithX, 0x94, 4,
+    tax_implied: TAX, Implied, 0xaa, 2,
+    tay_implied: TAY, Implied, 0xa8, 2,
+    tsx_implied: TSX, Implied, 0xba, 2,
+    txa_implied: TXA, Implied, 0x8a, 2,
+    txs_implied: TXS, Implied, 0x9a, 2,
+    tya_implied: TYA, Implied, 0x98, 2,
+    brk_implied: BRK, Implied, 0x00, 7,
+    nop_implied: NOP, Implied, 0xea, 2,
 );
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ADC,
-    addressing_mode::AbsoluteIndexedWithY,
-    0x79,
-    4
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::ADC, addressing_mode::IndirectYIndexed, 0x71, 5);
-
-gen_instruction_cycles_and_parser!(mnemonic::ADC, addressing_mode::Immediate, 0x69, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::ADC, addressing_mode::XIndexedIndirect, 0x61, 6);
-
-gen_instruction_cycles_and_parser!(mnemonic::ADC, addressing_mode::ZeroPage, 0x65, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ADC,
-    addressing_mode::ZeroPageIndexedWithX,
-    0x75,
-    4
-);
-
-// SBC
-
-gen_instruction_cycles_and_parser!(mnemonic::SBC, addressing_mode::Absolute, 0xed, 4);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::SBC,
-    addressing_mode::AbsoluteIndexedWithX,
-    0xFD,
-    4
-);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::SBC,
-    addressing_mode::AbsoluteIndexedWithY,
-    0xF9,
-    4
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::SBC, addressing_mode::IndirectYIndexed, 0xf1, 5);
-
-gen_instruction_cycles_and_parser!(mnemonic::SBC, addressing_mode::Immediate, 0xe9, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::SBC, addressing_mode::XIndexedIndirect, 0xe1, 6);
-
-gen_instruction_cycles_and_parser!(mnemonic::SBC, addressing_mode::ZeroPage, 0xe5, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::SBC,
-    addressing_mode::ZeroPageIndexedWithX,
-    0xf5,
-    4
-);
-
-// Bit-wise Operations
-
-// AND
-
-gen_instruction_cycles_and_parser!(mnemonic::AND, addressing_mode::Absolute, 0x2d, 4);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::AND,
-    addressing_mode::AbsoluteIndexedWithX,
-    0x3d,
-    4
-);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::AND,
-    addressing_mode::AbsoluteIndexedWithY,
-    0x39,
-    4
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::AND, addressing_mode::IndirectYIndexed, 0x31, 5);
-
-gen_instruction_cycles_and_parser!(mnemonic::AND, addressing_mode::Immediate, 0x29, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::AND, addressing_mode::XIndexedIndirect, 0x21, 6);
-
-gen_instruction_cycles_and_parser!(mnemonic::AND, addressing_mode::ZeroPage, 0x25, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::AND,
-    addressing_mode::ZeroPageIndexedWithX,
-    0x35,
-    4
-);
-
-// ASL
-
-gen_instruction_cycles_and_parser!(mnemonic::ASL, addressing_mode::Absolute, 0x0e, 6);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ASL,
-    addressing_mode::AbsoluteIndexedWithX,
-    0x1e,
-    7
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::ASL, addressing_mode::Accumulator, 0x0a, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::ASL, addressing_mode::ZeroPage, 0x06, 5);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ASL,
-    addressing_mode::ZeroPageIndexedWithX,
-    0x16,
-    6
-);
-
-// BIT
-
-gen_instruction_cycles_and_parser!(mnemonic::BIT, addressing_mode::Absolute, 0x2c, 4);
-
-gen_instruction_cycles_and_parser!(mnemonic::BIT, addressing_mode::ZeroPage, 0x24, 3);
-
-// EOR
-
-gen_instruction_cycles_and_parser!(mnemonic::EOR, addressing_mode::Absolute, 0x4d, 4);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::EOR,
-    addressing_mode::AbsoluteIndexedWithX,
-    0x5d,
-    4
-);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::EOR,
-    addressing_mode::AbsoluteIndexedWithY,
-    0x59,
-    4
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::EOR, addressing_mode::IndirectYIndexed, 0x51, 5);
-
-gen_instruction_cycles_and_parser!(mnemonic::EOR, addressing_mode::Immediate, 0x49, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::EOR, addressing_mode::XIndexedIndirect, 0x41, 6);
-
-gen_instruction_cycles_and_parser!(mnemonic::EOR, addressing_mode::ZeroPage, 0x45, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::EOR,
-    addressing_mode::ZeroPageIndexedWithX,
-    0x55,
-    4
-);
-
-// LSR
-
-gen_instruction_cycles_and_parser!(mnemonic::LSR, addressing_mode::Absolute, 0x4e, 6);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::LSR,
-    addressing_mode::AbsoluteIndexedWithX,
-    0x5e,
-    7
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::LSR, addressing_mode::Accumulator, 0x4a, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::LSR, addressing_mode::ZeroPage, 0x46, 5);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::LSR,
-    addressing_mode::ZeroPageIndexedWithX,
-    0x56,
-    6
-);
-
-// ORA
-
-gen_instruction_cycles_and_parser!(mnemonic::ORA, addressing_mode::Absolute, 0x0d, 4);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ORA,
-    addressing_mode::AbsoluteIndexedWithX,
-    0x1d,
-    4
-);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ORA,
-    addressing_mode::AbsoluteIndexedWithY,
-    0x19,
-    4
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::ORA, addressing_mode::IndirectYIndexed, 0x11, 5);
-
-gen_instruction_cycles_and_parser!(mnemonic::ORA, addressing_mode::Immediate, 0x09, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::ORA, addressing_mode::XIndexedIndirect, 0x01, 6);
-
-gen_instruction_cycles_and_parser!(mnemonic::ORA, addressing_mode::ZeroPage, 0x05, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ORA,
-    addressing_mode::ZeroPageIndexedWithX,
-    0x15,
-    4
-);
-
-// ROL
-
-gen_instruction_cycles_and_parser!(mnemonic::ROL, addressing_mode::Absolute, 0x2e, 6);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ROL,
-    addressing_mode::AbsoluteIndexedWithX,
-    0x3e,
-    7
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::ROL, addressing_mode::Accumulator, 0x2a, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::ROL, addressing_mode::ZeroPage, 0x26, 5);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ROL,
-    addressing_mode::ZeroPageIndexedWithX,
-    0x36,
-    6
-);
-
-// ROR
-
-gen_instruction_cycles_and_parser!(mnemonic::ROR, addressing_mode::Absolute, 0x6e, 6);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ROR,
-    addressing_mode::AbsoluteIndexedWithX,
-    0x7e,
-    7
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::ROR, addressing_mode::Accumulator, 0x6a, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::ROR, addressing_mode::ZeroPage, 0x66, 5);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::ROR,
-    addressing_mode::ZeroPageIndexedWithX,
-    0x76,
-    6
-);
-
-// Branching
-
-// BCC
-
-gen_instruction_cycles_and_parser!(mnemonic::BCC, addressing_mode::Relative, 0x90, 2);
-
-// BCS
-
-gen_instruction_cycles_and_parser!(mnemonic::BCS, addressing_mode::Relative, 0xb0, 2);
-
-// BEQ
-
-gen_instruction_cycles_and_parser!(mnemonic::BEQ, addressing_mode::Relative, 0xf0, 2);
-
-// BMI
-
-gen_instruction_cycles_and_parser!(mnemonic::BMI, addressing_mode::Relative, 0x30, 2);
-
-// BNE
-
-gen_instruction_cycles_and_parser!(mnemonic::BNE, addressing_mode::Relative, 0xd0, 2);
-
-// BPL
-
-gen_instruction_cycles_and_parser!(mnemonic::BPL, addressing_mode::Relative, 0x10, 2);
-
-// BVC
-
-gen_instruction_cycles_and_parser!(mnemonic::BVC, addressing_mode::Relative, 0x50, 2);
-
-// BVS
-
-gen_instruction_cycles_and_parser!(mnemonic::BVS, addressing_mode::Relative, 0x70, 2);
-
-// CLC
-
-gen_instruction_cycles_and_parser!(mnemonic::CLC, addressing_mode::Implied, 0x18, 2);
-
-// CLD
-
-gen_instruction_cycles_and_parser!(mnemonic::CLD, addressing_mode::Implied, 0xd8, 2);
-
-// CLI
-
-gen_instruction_cycles_and_parser!(mnemonic::CLI, addressing_mode::Implied, 0x58, 2);
-
-// CLV
-
-gen_instruction_cycles_and_parser!(mnemonic::CLV, addressing_mode::Implied, 0xb8, 2);
-
-// CMP
-
-gen_instruction_cycles_and_parser!(mnemonic::CMP, addressing_mode::Absolute, 0xcd, 4);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::CMP,
-    addressing_mode::AbsoluteIndexedWithX,
-    0xdd,
-    4
-);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::CMP,
-    addressing_mode::AbsoluteIndexedWithY,
-    0xd9,
-    4
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::CMP, addressing_mode::IndirectYIndexed, 0xd1, 5);
-
-gen_instruction_cycles_and_parser!(mnemonic::CMP, addressing_mode::Immediate, 0xc9, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::CMP, addressing_mode::XIndexedIndirect, 0xc1, 6);
-
-gen_instruction_cycles_and_parser!(mnemonic::CMP, addressing_mode::ZeroPage, 0xc5, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::CMP,
-    addressing_mode::ZeroPageIndexedWithX,
-    0xd5,
-    4
-);
-
-// CPX
-
-gen_instruction_cycles_and_parser!(mnemonic::CPX, addressing_mode::Absolute, 0xec, 4);
-
-gen_instruction_cycles_and_parser!(mnemonic::CPX, addressing_mode::Immediate, 0xe0, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::CPX, addressing_mode::ZeroPage, 0xe4, 3);
-
-// CPY
-
-gen_instruction_cycles_and_parser!(mnemonic::CPY, addressing_mode::Absolute, 0xcc, 4);
-
-gen_instruction_cycles_and_parser!(mnemonic::CPY, addressing_mode::Immediate, 0xc0, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::CPY, addressing_mode::ZeroPage, 0xc4, 3);
-
-// DEC
-
-gen_instruction_cycles_and_parser!(mnemonic::DEC, addressing_mode::Absolute, 0xce, 6);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::DEC,
-    addressing_mode::AbsoluteIndexedWithX,
-    0xde,
-    7
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::DEC, addressing_mode::ZeroPage, 0xc6, 5);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::DEC,
-    addressing_mode::ZeroPageIndexedWithX,
-    0xd6,
-    6
-);
-
-// DEX
-
-gen_instruction_cycles_and_parser!(mnemonic::DEX, addressing_mode::Implied, 0xca, 2);
-
-// DEY
-
-gen_instruction_cycles_and_parser!(mnemonic::DEY, addressing_mode::Implied, 0x88, 2);
-
-// INC
-
-gen_instruction_cycles_and_parser!(mnemonic::INC, addressing_mode::Absolute, 0xee, 6);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::INC,
-    addressing_mode::AbsoluteIndexedWithX,
-    0xfe,
-    7
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::INC, addressing_mode::ZeroPage, 0xe6, 5);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::INC,
-    addressing_mode::ZeroPageIndexedWithX,
-    0xf6,
-    6
-);
-
-// INX
-
-gen_instruction_cycles_and_parser!(mnemonic::INX, addressing_mode::Implied, 0xe8, 2);
-
-// INY
-
-gen_instruction_cycles_and_parser!(mnemonic::INY, addressing_mode::Implied, 0xc8, 2);
-
-// JMP
-
-gen_instruction_cycles_and_parser!(mnemonic::JMP, addressing_mode::Absolute, 0x4c, 3);
-
-gen_instruction_cycles_and_parser!(mnemonic::JMP, addressing_mode::Indirect, 0x6c, 5);
-
-// JSR
-
-gen_instruction_cycles_and_parser!(mnemonic::JSR, addressing_mode::Absolute, 0x20, 6);
-
-// LDA
-
-gen_instruction_cycles_and_parser!(mnemonic::LDA, addressing_mode::Immediate, 0xa9, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::LDA, addressing_mode::ZeroPage, 0xa5, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::LDA,
-    addressing_mode::ZeroPageIndexedWithX,
-    0xb5,
-    4
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::LDA, addressing_mode::Absolute, 0xad, 4);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::LDA,
-    addressing_mode::AbsoluteIndexedWithX,
-    0xbd,
-    4
-);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::LDA,
-    addressing_mode::AbsoluteIndexedWithY,
-    0xb9,
-    4
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::LDA, addressing_mode::IndirectYIndexed, 0xb1, 5);
-
-gen_instruction_cycles_and_parser!(mnemonic::LDA, addressing_mode::XIndexedIndirect, 0xa1, 6);
-
-// LDX
-
-gen_instruction_cycles_and_parser!(mnemonic::LDX, addressing_mode::Absolute, 0xae, 4);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::LDX,
-    addressing_mode::AbsoluteIndexedWithY,
-    0xbe,
-    4
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::LDX, addressing_mode::Immediate, 0xa2, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::LDX, addressing_mode::ZeroPage, 0xa6, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::LDX,
-    addressing_mode::ZeroPageIndexedWithY,
-    0xb6,
-    4
-);
-
-// LDY
-
-gen_instruction_cycles_and_parser!(mnemonic::LDY, addressing_mode::Absolute, 0xac, 4);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::LDY,
-    addressing_mode::AbsoluteIndexedWithX,
-    0xbc,
-    4
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::LDY, addressing_mode::Immediate, 0xa0, 2);
-
-gen_instruction_cycles_and_parser!(mnemonic::LDY, addressing_mode::ZeroPage, 0xa4, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::LDY,
-    addressing_mode::ZeroPageIndexedWithX,
-    0xb4,
-    4
-);
-
-// PHA
-
-gen_instruction_cycles_and_parser!(mnemonic::PHA, addressing_mode::Implied, 0x48, 3);
-
-// PHP
-
-gen_instruction_cycles_and_parser!(mnemonic::PHP, addressing_mode::Implied, 0x08, 3);
-
-// PLA
-
-gen_instruction_cycles_and_parser!(mnemonic::PLA, addressing_mode::Implied, 0x68, 4);
-
-// PLP
-
-gen_instruction_cycles_and_parser!(mnemonic::PLP, addressing_mode::Implied, 0x28, 4);
-
-// RTI
-
-gen_instruction_cycles_and_parser!(mnemonic::RTI, addressing_mode::Implied, 0x40, 6);
-
-// RTS
-
-gen_instruction_cycles_and_parser!(mnemonic::RTS, addressing_mode::Implied, 0x60, 6);
-
-// SEC
-
-gen_instruction_cycles_and_parser!(mnemonic::SEC, addressing_mode::Implied, 0x38, 2);
-
-// SED
-
-gen_instruction_cycles_and_parser!(mnemonic::SED, addressing_mode::Implied, 0xf8, 2);
-
-// SEI
-
-gen_instruction_cycles_and_parser!(mnemonic::SEI, addressing_mode::Implied, 0x78, 2);
-
-// STA
-
-gen_instruction_cycles_and_parser!(mnemonic::STA, addressing_mode::Absolute, 0x8d, 4);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::STA,
-    addressing_mode::AbsoluteIndexedWithX,
-    0x9d,
-    5
-);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::STA,
-    addressing_mode::AbsoluteIndexedWithY,
-    0x99,
-    5
-);
-
-gen_instruction_cycles_and_parser!(mnemonic::STA, addressing_mode::IndirectYIndexed, 0x91, 6);
-
-gen_instruction_cycles_and_parser!(mnemonic::STA, addressing_mode::XIndexedIndirect, 0x81, 6);
-
-gen_instruction_cycles_and_parser!(mnemonic::STA, addressing_mode::ZeroPage, 0x85, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::STA,
-    addressing_mode::ZeroPageIndexedWithX,
-    0x95,
-    4
-);
-
-// STX
-
-gen_instruction_cycles_and_parser!(mnemonic::STX, addressing_mode::Absolute, 0x8e, 4);
-
-gen_instruction_cycles_and_parser!(mnemonic::STX, addressing_mode::ZeroPage, 0x86, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::STX,
-    addressing_mode::ZeroPageIndexedWithY,
-    0x96,
-    4
-);
-
-// STY
-
-gen_instruction_cycles_and_parser!(mnemonic::STY, addressing_mode::Absolute, 0x8c, 4);
-
-gen_instruction_cycles_and_parser!(mnemonic::STY, addressing_mode::ZeroPage, 0x84, 3);
-
-gen_instruction_cycles_and_parser!(
-    mnemonic::STY,
-    addressing_mode::ZeroPageIndexedWithX,
-    0x94,
-    4
-);
-
-// TAX
-
-gen_instruction_cycles_and_parser!(mnemonic::TAX, addressing_mode::Implied, 0xaa, 2);
-
-// TAY
-
-gen_instruction_cycles_and_parser!(mnemonic::TAY, addressing_mode::Implied, 0xa8, 2);
-
-// TSX
-
-gen_instruction_cycles_and_parser!(mnemonic::TSX, addressing_mode::Implied, 0xba, 2);
-
-// TXA
-
-gen_instruction_cycles_and_parser!(mnemonic::TXA, addressing_mode::Implied, 0x8a, 2);
-
-// TSX
-
-gen_instruction_cycles_and_parser!(mnemonic::TXS, addressing_mode::Implied, 0x9a, 2);
-
-// TYA
-
-gen_instruction_cycles_and_parser!(mnemonic::TYA, addressing_mode::Implied, 0x98, 2);
-
-// Misc
-
-// BRK
-
-gen_instruction_cycles_and_parser!(mnemonic::BRK, addressing_mode::Implied, 0x00, 7);
-
-// NOP
-
-gen_instruction_cycles_and_parser!(mnemonic::NOP, addressing_mode::Implied, 0xea, 2);
